Add Extractor trait
Extractor is an entity that connects to SBT server and watches for some
keys. Extractor has the notion of step in importing process. Splitting
project retrieval into steps is necessary because it takes too
long to retrieve some keys (e.g. 'update' or 'updateClassifiers') and we
want first-time import to be fast.