Fix SbtConnector.open method / Reuse connectors
Found interesting bug: on Mac for each SbtConnector's instance a new JVM for
sbt-server is spawned whilst on Linux existing JVM is reused. Decided to fix it
by reusing SbtConnector's instance for particular path, but hit another
error: `open` method can not be called more than once. The source of this
behaviour is in calling `claimMessages` in SimpleSbtClient ctor. While the
reason of this is still unclear for me I've found a workaround for this in
reusing SbtClient instances as well.

NOTE: check on Windows. Sure there'll be a lot of unpleasant surpirses
NOTE: tell SBT guys about this behaviour